kvccl: move followerreadsccl under kvccl
Package movement.
