+++ pkg/ccl/kvccl/kvccl.go
@@ -8,4 +8,8 @@
 
 package kvccl
 
-import _ "github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvtenantccl" // init hooks
+import (
+	// ccl init hooks.
+	_ "github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvfollowerreadsccl"
+	_ "github.com/cockroachdb/cockroach/pkg/ccl/kvccl/kvtenantccl"
+)
